fix offergrouping and move it
